requirements.txt
toml
rich
click
schedule
appdirs
loguru
faker
tinydb
aiohttp
aiohttp_socks
python-dateutil
requests[socks]
ddddocr
embypy
python-telegram-client
python-telegram-bot[socks, job-queue]
